requirements.txt
cython
scipy
shapely
timm
h5py
submitit
scikit-image
wandb
setuptools
numpy==1.22.4
Pillow==9.3.0
pycocotools~=2.0.4
fvcore
tabulate
tqdm
ftfy
regex
opencv-python
<<<<<<< HEAD
open_clip_torch==2.16.0
=======
open_clip_torch==2.16.0
einops
git+https://github.com/facebookresearch/detectron2.git
>>>>>>> feabc4013b61fef153ad34f36e70956f1345eb0d
